+++ packages/fal/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ai-sdk/fal
 
+## 2.0.0-beta.40
+
+### Patch Changes
+
+- Updated dependencies [954c356]
+  - @ai-sdk/provider-utils@4.0.0-beta.39
+  - @ai-sdk/provider@3.0.0-beta.21
+
+## 2.0.0-beta.39
+
+### Patch Changes
+
+- Updated dependencies [03849b0]
+  - @ai-sdk/provider-utils@4.0.0-beta.38
+
 ## 2.0.0-beta.38
 
 ### Patch Changes