+++ packages/gateway/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @ai-sdk/gateway
 
+## 2.0.0-beta.67
+
+### Patch Changes
+
+- Updated dependencies [954c356]
+  - @ai-sdk/provider-utils@4.0.0-beta.39
+  - @ai-sdk/provider@3.0.0-beta.21
+
+## 2.0.0-beta.66
+
+### Patch Changes
+
+- Updated dependencies [03849b0]
+  - @ai-sdk/provider-utils@4.0.0-beta.38
+
+## 2.0.0-beta.65
+
+### Patch Changes
+
+- cdd0bc2: feat (provider/gateway): add intellect-3 model id
+
 ## 2.0.0-beta.64
 
 ### Patch Changes