+++ packages/google/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ai-sdk/google
 
+## 3.0.0-beta.61
+
+### Patch Changes
+
+- 954c356: feat(openai): allow custom names for provider-defined tools
+- Updated dependencies [954c356]
+  - @ai-sdk/provider-utils@4.0.0-beta.39
+  - @ai-sdk/provider@3.0.0-beta.21
+
+## 3.0.0-beta.60
+
+### Patch Changes
+
+- Updated dependencies [03849b0]
+  - @ai-sdk/provider-utils@4.0.0-beta.38
+
 ## 3.0.0-beta.59
 
 ### Patch Changes