+++ packages/llamaindex/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @ai-sdk/llamaindex
 
+## 2.0.0-beta.123
+
+### Patch Changes
+
+- Updated dependencies [954c356]
+  - ai@6.0.0-beta.123
+
+## 2.0.0-beta.122
+
+### Patch Changes
+
+- Updated dependencies [03849b0]
+  - ai@6.0.0-beta.122
+
+## 2.0.0-beta.121
+
+### Patch Changes
+
+- ai@6.0.0-beta.121
+
 ## 2.0.0-beta.120
 
 ### Patch Changes