+++ packages/mcp/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ai-sdk/mcp
 
+## 1.0.0-beta.23
+
+### Patch Changes
+
+- Updated dependencies [954c356]
+  - @ai-sdk/provider-utils@4.0.0-beta.39
+  - @ai-sdk/provider@3.0.0-beta.21
+
+## 1.0.0-beta.22
+
+### Patch Changes
+
+- Updated dependencies [03849b0]
+  - @ai-sdk/provider-utils@4.0.0-beta.38
+
 ## 1.0.0-beta.21
 
 ### Patch Changes