+++ packages/provider-utils/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/provider-utils
 
+## 4.0.0-beta.39
+
+### Patch Changes
+
+- 954c356: feat(openai): allow custom names for provider-defined tools
+- Updated dependencies [954c356]
+  - @ai-sdk/provider@3.0.0-beta.21
+
+## 4.0.0-beta.38
+
+### Patch Changes
+
+- 03849b0: move DelayedPromise into provider utils
+
 ## 4.0.0-beta.37
 
 ### Patch Changes