+++ packages/rsc/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @ai-sdk/rsc
 
+## 2.0.0-beta.124
+
+### Patch Changes
+
+- Updated dependencies [954c356]
+  - @ai-sdk/provider-utils@4.0.0-beta.39
+  - @ai-sdk/provider@3.0.0-beta.21
+  - ai@6.0.0-beta.123
+
+## 2.0.0-beta.123
+
+### Patch Changes
+
+- Updated dependencies [03849b0]
+  - @ai-sdk/provider-utils@4.0.0-beta.38
+  - ai@6.0.0-beta.122
+
+## 2.0.0-beta.122
+
+### Patch Changes
+
+- ai@6.0.0-beta.121
+
 ## 2.0.0-beta.121
 
 ### Patch Changes