+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,26 @@
 
 ### Patch Changes
 
+- Updated dependencies [954c356]
+  - ai@6.0.0-beta.123
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [03849b0]
+  - ai@6.0.0-beta.122
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@6.0.0-beta.121
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
 - Updated dependencies [457318b]
   - ai@6.0.0-beta.120