+++ packages/valibot/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/valibot
 
+## 2.0.0-beta.39
+
+### Patch Changes
+
+- Updated dependencies [954c356]
+  - @ai-sdk/provider-utils@4.0.0-beta.39
+
+## 2.0.0-beta.38
+
+### Patch Changes
+
+- Updated dependencies [03849b0]
+  - @ai-sdk/provider-utils@4.0.0-beta.38
+
 ## 2.0.0-beta.37
 
 ### Patch Changes